+++ CHANGELOG.md
@@ -14,6 +14,14 @@ For the purpose of determining breaking changes:
 
 [python-versions]: https://devguide.python.org/versions/#supported-versions
 
+## [Unreleased]
+
+[unreleased]: https://github.com/rogdham/bigxml/compare/v1.0.1...HEAD
+
+### :rocket: Added
+
+- Add support for buffer protocol ([PEP 688](https://peps.python.org/pep-0688/))
+
 ## [1.0.1] - 2024-04-27
 
 [1.0.1]: https://github.com/rogdham/bigxml/compare/v1.0.0...v1.0.1